# src/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IGenericFailureResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEmailBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"consult": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"phone",
					"website",
					"consult",
					"comment"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "digilance-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/landing-page/data": {
			"get": {
				"operationId": "GetLandingPageData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"$ref": "#/components/schemas/IGenericFailureResponse"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/landing-page/email": {
			"post": {
				"operationId": "SendEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmailBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}